refactor(runtime): generalize Vercel sandbox into a vault-backed RuntimeProvider abstraction
Reshapes #1585 so the worker bash backend is a pluggable runtime provider
instead of a Vercel-hardwired flag + route + env-var credentials.

- worker: RuntimeProvider registry + one generic /internal/runtime/exec client;
  selection via LOBU_RUNTIME_PROVIDER (renamed from LOBU_WORKSPACE_BACKEND)
- gateway: GatewayRuntimeProvider registry + generic route; provider chosen from
  the signed worker-token claim (worker never names a provider over the wire)
- credentials: resolved gateway-side per-environment from the vault
  (environment:<id>:<field>) with system-env fallback; worker never sees them
- token: runtimeProviderId + environmentId claims, validated in verifyWorkerToken
- Vercel becomes one provider declaration on each side; old vercel-sandbox files removed
- fixes a merge-surfaced http-proxy checkDomainAccess arity break

Tests: generic-runtime-bash 3/3, runtime-route 10/10, embedded-deployment 31/31.
