+++ .env.example
@@ -92,13 +92,15 @@ WORKER_ALLOWED_DOMAINS=api.anthropic.com,.anthropic.com,registry.npmjs.org,.npmj
 # Explicitly blocked domains (optional, use with WORKER_ALLOWED_DOMAINS=*)
 WORKER_DISALLOWED_DOMAINS=
 
-# Optional remote workspace backend for worker bash commands.
+# Optional remote runtime provider for worker bash commands (deployment-wide
+# default / self-host; per-agent selection comes from its Environment).
 # Default/unset: local per-conversation workspaces with embedded just-bash.
 # vercel: route bash through a Vercel persistent sandbox. The sandbox filesystem
 # is the workspace source of truth; regular files are not synced locally.
-# Requires Vercel Sandbox credentials in the app process environment:
+# Provider credentials are read from the org vault (per-environment) and fall
+# back to the app process environment here:
 # VERCEL_TEAM_ID + VERCEL_PROJECT_ID + VERCEL_TOKEN.
-# LOBU_WORKSPACE_BACKEND=vercel
+# LOBU_RUNTIME_PROVIDER=vercel
 # LOBU_VERCEL_SANDBOX_NAME_PREFIX=lobu
 # LOBU_VERCEL_SANDBOX_RUNTIME=node24
 # VERCEL_SANDBOX_DEFAULT_RUNTIME=node24
