+++ packages/agent-worker/src/embedded/just-bash-bootstrap.ts
@@ -23,10 +23,7 @@ import {
 } from "./exec-sandbox";
 import type { McpCliCommand, McpRuntimeRef } from "./mcp-cli-commands";
 import { buildMcpCliCommands } from "./mcp-cli-commands";
-import {
-  createVercelSandboxBashOps,
-  useVercelSandboxBackend,
-} from "./vercel-sandbox-bash";
+import { getWorkerRuntimeProvider } from "./runtime";
 
 const EMBEDDED_BASH_LIMITS = {
   maxCommandCount: 50_000,
@@ -407,19 +404,24 @@ async function adaptMcpCliCommand(
 export async function createEmbeddedBashOps(
   options: EmbeddedBashOpsOptions = {}
 ): Promise<BashOperations> {
-  if (useVercelSandboxBackend()) {
+  const runtimeProvider = getWorkerRuntimeProvider(
+    process.env.LOBU_RUNTIME_PROVIDER
+  );
+  if (runtimeProvider) {
     if (!options.gw) {
       throw new Error(
-        "LOBU_WORKSPACE_BACKEND=vercel requires gateway parameters"
+        `LOBU_RUNTIME_PROVIDER=${runtimeProvider.id} requires gateway parameters`
       );
     }
     if (options.mcpExposure === "cli") {
       console.warn(
-        "[embedded] LOBU_WORKSPACE_BACKEND=vercel routes bash through Vercel; MCP CLI exposure is unavailable there. Use MCP tools exposure instead."
+        `[embedded] LOBU_RUNTIME_PROVIDER=${runtimeProvider.id} routes bash through a remote runtime; MCP CLI exposure is unavailable there. Use MCP tools exposure instead.`
       );
     }
-    console.log("[embedded] bash backend active: vercel persistent sandbox");
-    return createVercelSandboxBashOps({ gw: options.gw });
+    console.log(
+      `[embedded] bash backend active: ${runtimeProvider.id} runtime provider`
+    );
+    return runtimeProvider.createBashOps({ gw: options.gw });
   }
 
   const { Bash, ReadWriteFs } = await import("just-bash");
