+++ packages/agent-worker/src/embedded/runtime/generic-runtime-bash.ts
@@ -1,9 +1,10 @@
 import { stripEnv } from "@lobu/core";
 import type { BashOperations } from "@mariozechner/pi-coding-agent";
-import type { GatewayParams } from "../shared/tool-implementations";
-import { SENSITIVE_WORKER_ENV_KEYS } from "../shared/worker-env-keys";
+import type { GatewayParams } from "../../shared/tool-implementations";
+import { SENSITIVE_WORKER_ENV_KEYS } from "../../shared/worker-env-keys";
+import type { WorkerRuntimeProvider } from "./types";
 
-type VercelSandboxExecResponse = {
+type RuntimeExecResponse = {
   stdout?: unknown;
   stderr?: unknown;
   exitCode?: unknown;
@@ -11,6 +12,13 @@ type VercelSandboxExecResponse = {
 };
 
 const DOMAIN_PATTERN = /^[A-Za-z0-9.*_-]+(?::\d+)?$/;
+
+/**
+ * The worker egresses through a local gateway HTTP proxy (`HTTP_PROXY=…:8118`),
+ * which is meaningless inside a remote sandbox — the sandbox enforces egress
+ * via its own network policy derived from `allowedDomains`. Strip them so the
+ * remote command doesn't try to dial a proxy that isn't there.
+ */
 const REMOTE_UNSUPPORTED_ENV_KEYS = [
   "ALL_PROXY",
   "HTTPS_PROXY",
@@ -38,26 +46,27 @@ function parseAllowedDomains(): string[] {
 }
 
 function commandEnv(
-  env: NodeJS.ProcessEnv | undefined
+  env: NodeJS.ProcessEnv | undefined,
+  remoteEnv: Record<string, string>
 ): Record<string, string> {
   const cleanEnv = stripEnv(env ?? process.env, [
     ...SENSITIVE_WORKER_ENV_KEYS,
     ...REMOTE_UNSUPPORTED_ENV_KEYS,
   ]);
-  return {
-    ...cleanEnv,
-    HOME: "/vercel/sandbox",
-    TMPDIR: "/vercel/sandbox/.tmp",
-    TMP: "/vercel/sandbox/.tmp",
-    TEMP: "/vercel/sandbox/.tmp",
-    XDG_CACHE_HOME: "/vercel/sandbox/.cache",
-  };
+  return { ...cleanEnv, ...remoteEnv };
 }
 
-export function createVercelSandboxBashOps(params: {
-  gw: GatewayParams;
-}): BashOperations {
-  const endpoint = `${params.gw.gatewayUrl.replace(/\/+$/, "")}/internal/vercel-sandbox/exec`;
+/**
+ * The single worker-side client for every remote runtime provider. POSTs to
+ * the generic `/internal/runtime/exec` route with the worker token; the body
+ * never names a provider (the gateway reads it from the signed token). No
+ * streaming — the full JSON result is awaited, then emitted via `onData`.
+ */
+export function createGenericRuntimeBashOps(
+  provider: WorkerRuntimeProvider,
+  params: { gw: GatewayParams }
+): BashOperations {
+  const endpoint = `${params.gw.gatewayUrl.replace(/\/+$/, "")}/internal/runtime/exec`;
 
   return {
     async exec(command, cwd, { env, onData, signal, timeout }) {
@@ -74,7 +83,7 @@ export function createVercelSandboxBashOps(params: {
           cwd,
           workspaceDir: params.gw.workspaceDir,
           timeoutMs,
-          env: commandEnv(env),
+          env: commandEnv(env, provider.remoteEnv),
           allowedDomains: parseAllowedDomains(),
         }),
         signal,
@@ -82,13 +91,13 @@ export function createVercelSandboxBashOps(params: {
 
       const payload = (await response
         .json()
-        .catch(() => ({}))) as VercelSandboxExecResponse;
+        .catch(() => ({}))) as RuntimeExecResponse;
 
       if (!response.ok) {
         const message =
           typeof payload.error === "string"
             ? payload.error
-            : `Vercel sandbox exec failed with HTTP ${response.status}`;
+            : `Runtime exec failed with HTTP ${response.status}`;
         onData(Buffer.from(`${message}\n`));
         return { exitCode: 1 };
       }
@@ -106,11 +115,4 @@ export function createVercelSandboxBashOps(params: {
       };
     },
   };
-}
-
-export function useVercelSandboxBackend(
-  backend = process.env.LOBU_WORKSPACE_BACKEND
-): boolean {
-  const value = backend?.toLowerCase();
-  return value === "vercel" || value === "vercel-sandbox";
 }
