+++ packages/server/src/gateway/__tests__/embedded-deployment.test.ts
@@ -452,14 +452,14 @@ describe("DeploymentManager", () => {
       }
     });
 
-    test("forwards Vercel workspace backend selector without Vercel credentials", async () => {
+    test("forwards runtime provider selector without provider credentials", async () => {
       const previous = {
-        LOBU_WORKSPACE_BACKEND: process.env.LOBU_WORKSPACE_BACKEND,
+        LOBU_RUNTIME_PROVIDER: process.env.LOBU_RUNTIME_PROVIDER,
         VERCEL_TOKEN: process.env.VERCEL_TOKEN,
         VERCEL_TEAM_ID: process.env.VERCEL_TEAM_ID,
         VERCEL_PROJECT_ID: process.env.VERCEL_PROJECT_ID,
       };
-      process.env.LOBU_WORKSPACE_BACKEND = "vercel";
+      process.env.LOBU_RUNTIME_PROVIDER = "vercel";
       process.env.VERCEL_TOKEN = "vercel-test-token";
       process.env.VERCEL_TEAM_ID = "team_test";
       process.env.VERCEL_PROJECT_ID = "prj_test";
@@ -473,7 +473,7 @@ describe("DeploymentManager", () => {
           | { env?: Record<string, string> }
           | undefined;
 
-        expect(spawnOptions?.env?.LOBU_WORKSPACE_BACKEND).toBe("vercel");
+        expect(spawnOptions?.env?.LOBU_RUNTIME_PROVIDER).toBe("vercel");
         expect(spawnOptions?.env?.VERCEL_TOKEN).toBeUndefined();
         expect(spawnOptions?.env?.VERCEL_TEAM_ID).toBeUndefined();
         expect(spawnOptions?.env?.VERCEL_PROJECT_ID).toBeUndefined();
