+++ packages/server/src/gateway/orchestration/worker-token-claims.ts
@@ -19,6 +19,16 @@ export interface WorkerTokenClaimsArgs {
   organizationId?: string;
   platform?: string;
   platformMetadata?: Record<string, unknown>;
+  /**
+   * Selected runtime provider for this agent, resolved from its environment by
+   * the caller (which has the agent settings + environments store). When
+   * omitted, falls back to the deployment-wide `LOBU_RUNTIME_PROVIDER` (self-
+   * host / org default). The generic runtime route reads this claim to pick a
+   * provider — see WorkerTokenData.runtimeProviderId.
+   */
+  runtimeProviderId?: string;
+  /** The `environments.id` whose vault credential backs the provider above. */
+  environmentId?: string;
 }
 
 /**
@@ -43,7 +53,16 @@ export function buildWorkerTokenClaims(args: WorkerTokenClaimsArgs): {
   platform?: string;
   connectionId?: string;
   source?: string;
+  runtimeProviderId?: string;
+  environmentId?: string;
 } {
+  // Per-agent environment selection wins; otherwise the deployment-wide
+  // selector covers self-host / org-default. Resolving the env fallback here
+  // keeps both mints in lockstep (the same parity invariant as the routing
+  // claims above).
+  const runtimeProviderId =
+    args.runtimeProviderId ??
+    (process.env.LOBU_RUNTIME_PROVIDER?.trim() || undefined);
   return {
     channelId: args.channelId,
     teamId: args.teamId,
@@ -58,5 +77,10 @@ export function buildWorkerTokenClaims(args: WorkerTokenClaimsArgs): {
       typeof args.platformMetadata?.source === "string"
         ? args.platformMetadata.source
         : undefined,
+    runtimeProviderId,
+    // Only meaningful when a per-agent environment drove the selection; the
+    // env-var fallback has no environment row, so credentials resolve from
+    // system env.
+    environmentId: runtimeProviderId ? args.environmentId : undefined,
   };
 }
