+++ packages/server/src/gateway/runtime/index.ts
@@ -1,0 +1,16 @@
+/**
+ * Gateway-side runtime-provider registration barrel. Importing this registers
+ * every built-in provider as a side effect; adding a provider is one new file
+ * under `providers/` + one `registerGatewayRuntimeProvider` call here.
+ */
+import { vercelGatewayRuntimeProvider } from "./providers/vercel.js";
+import { registerGatewayRuntimeProvider } from "./registry.js";
+
+registerGatewayRuntimeProvider(vercelGatewayRuntimeProvider);
+
+export { getGatewayRuntimeProvider } from "./registry.js";
+export type {
+  GatewayRuntimeProvider,
+  RuntimeExecContext,
+  RuntimeExecResult,
+} from "./types.js";
